Created separate 'app' and 'extension'.
The app is a Chrome Packaged App, which is required to have chrome.bluetooth functionality. It doesn't need to have a window, so we'll just do it all within the scripts it runs.

Extension can't run chrome.bluetooth, so we use this as the front end for the user, and use cross extension messaging to get bluetooth data from the app.